+++ frontend/app/layout.tsx
@@ -1,13 +1,11 @@
-// import "./globals.css";
+'use client'; // Required for createTheme
 import { Inter } from "next/font/google";
 import { PublicEnvScript } from "next-runtime-env";
+import { ThemeProvider } from '@mui/material/styles';
+import CssBaseline from "@mui/material/CssBaseline";
+import theme from "./theme";
 
 const inter = Inter({ subsets: ["latin"] });
-
-export const metadata = {
-  title: "🛰 LHASA forecast",
-  description: "Current landslide predictions",
-};
 
 export default function RootLayout({
   children,
@@ -17,11 +15,16 @@ export default function RootLayout({
   return (
     <html lang="en">
       <head>
-        {/* see https://github.com/yurochka-dev/pet_project/blob/PTPJ7_docker_setup/frontend/src/app/layout.tsx */}
-        <PublicEnvScript />
+        <title>🛰 LHASA forecast</title>
+        <meta name="description" content="Current landslide predictions" />
+          {/* <PublicEnvScript /> */}
         <link rel="icon" href="/favicon.png" />
       </head>
-      <body className={inter.className}>{children}</body>
+      {/* Wrap the body content with ThemeProvider and CssBaseline */}
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <body className={inter.className}>{children}</body>
+      </ThemeProvider>
     </html>
   );
 }
