+++ frontend/app/theme.tsx
@@ -1,0 +1,19 @@
+'use client'; // Required for createTheme
+
+import { createTheme } from '@mui/material/styles';
+
+const theme = createTheme({
+  palette: {
+    mode: 'dark',
+    primary: {
+      main: '#ce93d8',
+    },
+    secondary: {
+      main: '#E0C2FF',
+      light: '#F5EBFF',
+      contrastText: '#47008F',
+    },
+  },
+});
+
+export default theme;
